+++ .vscode/settings.json
@@ -1,3 +1,13 @@
 {
-  "cSpell.words": ["txid", "txids", "Yubikey"]
+  "cSpell.words": [
+    "Caprese",
+    "FACTORINSTANCE",
+    "FACTORSOURCE",
+    "Quartier",
+    "Rémy",
+    "substate",
+    "txid",
+    "txids",
+    "Yubikey"
+  ]
 }
