+++ src/drivers/signing_drivers_context.rs
@@ -1,5 +1,5 @@
 use crate::prelude::*;
 
-pub trait IsSigningDriversContext {
+pub trait IsUseFactorSourceDriversContext {
     fn driver_for_factor_source_kind(&self, kind: FactorSourceKind) -> SigningDriver;
 }
