+++ src/lib.rs
@@ -4,6 +4,7 @@
 
 mod drivers;
 mod owned_types;
+mod petition_types;
 mod signatures_building_coordinator;
 mod signatures_outcome_types;
 mod testing;
@@ -12,6 +13,7 @@ mod types;
 pub mod prelude {
     pub use crate::drivers::*;
     pub use crate::owned_types::*;
+    pub use crate::petition_types::*;
     pub use crate::signatures_building_coordinator::*;
     pub use crate::signatures_outcome_types::*;
     pub use crate::testing::*;
