+++ src/petition_types/factor_list_kind.rs
@@ -1,0 +1,6 @@
+/// A kind of factor list, either threshold, or override kind.
+#[derive(PartialEq, Eq, Clone, Debug)]
+pub enum FactorListKind {
+    Threshold,
+    Override,
+}
