+++ src/signatures_building_coordinator.rs
@@ -1,4 +1,41 @@
+use std::marker::PhantomData;
+
 use crate::prelude::*;
+
+#[async_trait]
+pub trait DerivableFactorOutput {
+    type Input: std::hash::Hash;
+    async fn derive_from_source(
+        source: &FactorSource,
+        input: Self::Input,
+        factor_instance: OwnedFactorInstance,
+    ) -> Self;
+}
+
+pub trait IsOutputGroup {
+    type Output: DerivableFactorOutput;
+    fn input(&self) -> <Self::Output as DerivableFactorOutput>::Input;
+    fn supports_skipping_of_factors() -> bool;
+}
+
+/// A coordinator which gathers output from factor sources for each factor instance,
+/// associated with OutputGroups. An OutputGroup is either a `TransactionManifest`
+/// to sign or a `SecurityStructureOfFactorSources` to derive public keys with in
+/// order to create a `SecurityStructureOfFactorInstances`.
+///
+/// The coordinator supports building output from many factor sources for a batch
+/// of OutputGroups.
+///
+/// The user is prompted to use the factor sources in increasing friction order,
+/// for
+pub struct FactorSourcesOutputBuildingCoordinator<G: IsOutputGroup> {
+    phantom: PhantomData<G>,
+}
+impl<G: IsOutputGroup> FactorSourcesOutputBuildingCoordinator<G> {
+    pub fn new(_inputs: IndexSet<<G::Output as DerivableFactorOutput>::Input>) -> Self {
+        todo!()
+    }
+}
 
 /// A coordinator which gathers signatures from several factor sources of different
 /// kinds, in increasing friction order, for many transactions and for
@@ -9,30 +46,32 @@ use crate::prelude::*;
 /// is last; namely `DeviceFactorSource`, and the most tedious FactorSourceKind is
 /// first; namely `LedgerFactorSource`, which user might also lack access to.
 pub struct SignaturesBuildingCoordinator {
-    /// A context of drivers used to sign with factor sources.
-    signing_drivers_context: Arc<dyn IsSigningDriversContext>,
+    /// A context of drivers for "using" factor sources - either to (batch) sign
+    /// transaction(s) with, or to derive public keys from.
+    drivers: Arc<dyn IsUseFactorSourceDriversContext>,
 
-    /// Factor sources grouped by kind, sorted according to "signing order",
-    /// that is, we want to control which factor source kind users signs with
-    /// first, second etc, e.g. typically we prompt user to sign with Ledgers
+    /// Factor sources grouped by kind, sorted according to "friction order",
+    /// that is, we want to control which FactorSourceKind users use
+    /// first, second etc, e.g. typically we prompt user to use Ledgers
     /// first, and if a user might lack access to that Ledger device, then it is
     /// best to "fail fast", otherwise we might waste the users time, if she has
-    /// e.g. answered security questions and then is asked to sign with a Ledger
+    /// e.g. answered security questions and then is asked to use a Ledger
     /// she might not have handy at the moment - or might not be in front of a
     /// computer and thus unable to make a connection between the Radix Wallet
     /// and a Ledger device.
     factors_of_kind: IndexMap<FactorSourceKind, IndexSet<FactorSource>>,
 
-    /// Signature builders for each factor source, for each transaction,
-    /// for each entity to sign.
-    petitions: RefCell<Petitions>,
+    /// FactorSource output builders for each factor source, for each OutputGroup
+    /// (Transaction or SecurityStructureOfFactorInstances), for each entity to
+    /// produce an output (signature or public key).
+    builders: RefCell<Builders>,
 }
 
 impl SignaturesBuildingCoordinator {
     pub fn new(
         all_factor_sources_in_profile: IndexSet<FactorSource>,
         transactions: IndexSet<TransactionIntent>,
-        signing_drivers_context: Arc<dyn IsSigningDriversContext>,
+        signing_drivers_context: Arc<dyn IsUseFactorSourceDriversContext>,
     ) -> Self {
         let mut petitions_for_all_transactions =
             IndexMap::<IntentHash, PetitionOfTransaction>::new();
@@ -65,7 +104,7 @@ impl SignaturesBuildingCoordinator {
 
         for transaction in transactions.into_iter() {
             let mut petitions_for_entities =
-                HashMap::<AccountAddressOrIdentityAddress, PetitionOfTransactionByEntity>::new();
+                HashMap::<AccountAddressOrIdentityAddress, BuilderEntity>::new();
 
             for entity in transaction.clone().entities_requiring_auth {
                 let address = entity.address;
@@ -82,7 +121,7 @@ impl SignaturesBuildingCoordinator {
 
                         add(primary_role_matrix.override_factors.clone());
                         add(primary_role_matrix.threshold_factors.clone());
-                        let petition = PetitionOfTransactionByEntity::new_securified(
+                        let petition = BuilderEntity::new_securified(
                             transaction.intent_hash.clone(),
                             address.clone(),
                             primary_role_matrix,
@@ -93,7 +132,7 @@ impl SignaturesBuildingCoordinator {
                         let factor_instance = uec;
                         let factor_source_id = factor_instance.factor_source_id;
                         use_factor_in_tx(&factor_source_id, &transaction.intent_hash);
-                        let petition = PetitionOfTransactionByEntity::new_unsecurified(
+                        let petition = BuilderEntity::new_unsecurified(
                             transaction.intent_hash.clone(),
                             address.clone(),
                             factor_instance,
@@ -120,12 +159,12 @@ impl SignaturesBuildingCoordinator {
 
         factors_of_kind.sort_keys();
 
-        let petitions = Petitions::new(factor_to_payloads, petitions_for_all_transactions);
+        let petitions = Builders::new(factor_to_payloads, petitions_for_all_transactions);
 
         Self {
-            signing_drivers_context,
+            drivers: signing_drivers_context,
             factors_of_kind,
-            petitions: petitions.into(),
+            builders: petitions.into(),
         }
     }
 }
@@ -136,12 +175,11 @@ impl SignaturesBuildingCoordinator {
     ///
     /// `Ok(true)` means "continue", `Ok(false)` means "stop, we are done". `Err(_)` means "stop, we have failed".
     fn continue_if_necessary(&self) -> Result<bool> {
-        self.petitions.borrow().continue_if_necessary()
+        self.builders.borrow().continue_if_necessary()
     }
 
     fn get_driver(&self, kind: FactorSourceKind) -> SigningDriver {
-        self.signing_drivers_context
-            .driver_for_factor_source_kind(kind)
+        self.drivers.driver_for_factor_source_kind(kind)
     }
 
     async fn sign_with_factor_sources(
@@ -178,7 +216,7 @@ impl SignaturesBuildingCoordinator {
         &self,
         factor_source_id: &FactorSourceID,
     ) -> IndexMap<IntentHash, IndexSet<SerialSingleSigningRequestPartial>> {
-        self.petitions
+        self.builders
             .borrow()
             .inputs_for_serial_single_driver(factor_source_id)
     }
@@ -187,7 +225,7 @@ impl SignaturesBuildingCoordinator {
         &self,
         factor_source_id: &FactorSourceID,
     ) -> BatchTXBatchKeySigningRequest {
-        self.petitions
+        self.builders
             .borrow()
             .input_for_parallel_batch_driver(factor_source_id)
     }
@@ -196,7 +234,7 @@ impl SignaturesBuildingCoordinator {
         &self,
         factor_source_id: &FactorSourceID,
     ) -> IndexSet<InvalidTransactionIfSkipped> {
-        self.petitions
+        self.builders
             .borrow()
             .invalid_transactions_if_skipped(factor_source_id)
     }
@@ -217,7 +255,7 @@ impl SignaturesBuildingCoordinator {
         response: SignWithFactorSourceOrSourcesOutcome<HDSignature>,
     ) -> bool {
         {
-            let petitions = self.petitions.borrow_mut();
+            let petitions = self.builders.borrow_mut();
             petitions.process_single_response(response);
         }
         self.continue_if_necessary().unwrap_or(false)
@@ -226,7 +264,7 @@ impl SignaturesBuildingCoordinator {
         &self,
         response: SignWithFactorSourceOrSourcesOutcome<BatchSigningResponse>,
     ) {
-        let petitions = self.petitions.borrow_mut();
+        let petitions = self.builders.borrow_mut();
         petitions.process_batch_response(response)
     }
 }
@@ -234,6 +272,6 @@ impl SignaturesBuildingCoordinator {
 impl SignaturesBuildingCoordinator {
     pub async fn sign(self) -> SignaturesOutcome {
         self.do_sign().await;
-        self.petitions.into_inner().outcome()
+        self.builders.into_inner().outcome()
     }
 }
