+++ src/types/mod.rs
@@ -1,13 +1,11 @@
 mod hd_signature;
 mod hd_signature_input;
 mod invalid_transaction_if_skipped;
-mod petitions;
 mod sargon_types;
 mod sign_with_factor_source_or_sources_outcome;
 
 pub use hd_signature::*;
 pub use hd_signature_input::*;
 pub use invalid_transaction_if_skipped::*;
-pub use petitions::*;
 pub use sargon_types::*;
 pub use sign_with_factor_source_or_sources_outcome::*;
